refactor: 유틸함수 수정, 모달 컴포넌트 분리
