+++ week3_assignment/number-game/src/components/Modal/ModalContainer.jsx
@@ -1,0 +1,9 @@
+import { createPortal } from "react-dom";
+
+import { backdropStyle } from "./ModalContent.style";
+
+const ModalContainer = ({ children }) => {
+  return createPortal(<div css={backdropStyle}>{children}</div>, document.body);
+};
+
+export default ModalContainer;
